requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic[email]==2.3.0
email-validator==2.0.0
numpy==1.26.0
pandas==2.1.1
scikit-learn==1.3.1
xgboost==2.0.0
shap==0.42.1
faker==19.6.0
sqlalchemy==2.0.21
alembic==1.16.5
psycopg2-binary==2.9.9
asyncpg==0.28.0
redis==5.0.0
python-jose[cryptography]==3.3.0
statsmodels==0.14.0
python-multipart==0.0.6
matplotlib==3.8.0
httpx==0.27.2
